Created table User
