+++ backend/src/.env
@@ -1,0 +1,1 @@
+DATABASE_URL=sqlite+aiosqlite:///disChan.db
