+++ backend/src/alembic/versions/830d906f7101_.py
@@ -1,0 +1,42 @@
+"""empty message
+
+Revision ID: 830d906f7101
+Revises: 
+Create Date: 2023-10-20 21:28:00.239142
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '830d906f7101'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('user',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('username', sa.String(length=64), nullable=False),
+    sa.Column('email', sa.String(length=120), nullable=False),
+    sa.Column('hash_pass', sa.String(length=128), nullable=False),
+    sa.Column('salt', sa.String(length=1024), nullable=True),
+    sa.Column('about_me', sa.String(length=140), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('email'),
+    sa.UniqueConstraint('hash_pass')
+    )
+    op.create_index(op.f('ix_user_username'), 'user', ['username'], unique=True)
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_user_username'), table_name='user')
+    op.drop_table('user')
+    # ### end Alembic commands ###
